# Ensure the gather_mm is in order