First of hopefully a few that will increase MoE prompt processing speed. This is extremely simple and it comes with modest speedups but across the board so I think it is worth merging as a first step.

Some speed measurements:

| Model | Prompt | Before | After | Speedup |
| --- | --- | --- | --- | --- |
| DeepSeek-v3-0324 @ 4bit | ~1100 tokens | 78 tps | 112 tps | 43% |
| Mixtral-8x7B @ 4bit | ~1300 tokens | 174 tps | 192 tps | 10% |
| DeepSeek-V2-Lite-Chat @ 4bit | ~1200 tokens | 883 tps | 1033 tops | 17% |
| Qwen1.5-MoE-A2.7B @ 4bit | ~1200 tokens | 1200 tps | 1400 tps | 17% |