@ivanfioravanti 2025-03-28
DeepSeek-V2-Lite-Chat @ 4bit on M3 Ultra

Pre: Prompt: `4395 tokens, 1237.680 tokens-per-sec`
Post: Prompt: `4395 tokens, 1430.574 tokens-per-sec`

Thanks 🙏